Create new type of camera.
In order to facilitate easier testing file system camera is added. It allows to
specify a directory with a image files, which it will broadcast on the image
path, which the AI will pick up for processing.

Random camera is still a placeholder for future variations that will require
particular scenarios to be played.